requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
